requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
